Organise the conversion funnel files under one directory
